+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Packages:
 
 - [api-gateway/CHANGELOG.md](./packages/api-gateway/CHANGELOG.md)
 - [common-i18n/CHANGELOG.md](./packages/common-i18n/CHANGELOG.md)
-- [core-lib/CHANGELOG.md](./packages/core-lib/CHANGELOG.md)
+- [core/CHANGELOG.md](./packages/core/CHANGELOG.md)
 - [db-main-prisma/CHANGELOG.md](./packages/db-main-prisma/CHANGELOG.md)
-- [ui-lib/CHANGELOG.md](./packages/ui-lib/CHANGELOG.md)
+- [components/CHANGELOG.md](./packages/components/CHANGELOG.md)
 - [graphql-mesh/CHANGELOG.md](./packages/graphql-mesh/CHANGELOG.md)